Rename package to mavis.reporting
Move code from the old mavis_reporting namespace to the
new mavis.reporting layout.
