+++ .env.example
@@ -1,9 +1,9 @@
-FLASK_APP=mavis_reporting:create_app
+FLASK_APP=mavis.reporting:create_app
 FLASK_ENV=development
 # Required by Flask - used internally
 SECRET_KEY=(some random hex string)
-# Used as part of the OAuth 2.0 Authorization Code flow 
-# to identify this application to Mavis when requesting an 
+# Used as part of the OAuth 2.0 Authorization Code flow
+# to identify this application to Mavis when requesting an
 # authorization code
 CLIENT_ID=(some random hex string)
 # Used for signing / decoding-and-verifying-the-signature-of JWTs
