+++ mavis/reporting/assets/scss/_app_card.scss
