+++ mavis/reporting/assets/scss/_config.scss
