+++ mavis/reporting/assets/scss/_data_table.scss
