+++ mavis/reporting/helpers/static_file_helper.py
