+++ mavis/reporting/javascript/app.js
