+++ mavis/reporting/static/fonts/FrutigerLTW01-65Bold.woff2
