+++ mavis/reporting/templates/components/filters.jinja
