+++ mavis/reporting/templates/components/secondary_navigation.jinja
