+++ mavis/reporting/templates/components/table.jinja
