+++ mavis/reporting/templates/default.jinja
