+++ mavis/reporting/templates/download.jinja
