+++ mavis/reporting/templates/layouts/base.jinja
