+++ mavis/reporting/templates/layouts/dashboard.jinja
