+++ package.json
@@ -3,10 +3,10 @@
   "type": "module",
   "private": true,
   "scripts": {
-    "build:scss": "sass --load-path=node_modules mavis_reporting/assets/scss/app.scss:mavis_reporting/static/css/app.css --style=compressed --no-source-map",
-    "build:scss:dev": "sass --load-path=node_modules --watch mavis_reporting/assets/scss:mavis_reporting/static/css --style=expanded --source-map",
-    "build:js": "esbuild mavis_reporting/javascript/app.js --bundle --outfile=mavis_reporting/static/js/app.js --minify",
-    "build:js:dev": "esbuild mavis_reporting/javascript/app.js --bundle --outfile=mavis_reporting/static/js/app.js --sourcemap --minify --watch"
+    "build:scss": "sass --load-path=node_modules mavis/reporting/assets/scss/app.scss:mavis/reporting/static/css/app.css --style=compressed --no-source-map",
+    "build:scss:dev": "sass --load-path=node_modules --watch mavis/reporting/assets/scss:mavis/reporting/static/css --style=expanded --source-map",
+    "build:js": "esbuild mavis/reporting/javascript/app.js --bundle --outfile=mavis/reporting/static/js/app.js --minify",
+    "build:js:dev": "esbuild mavis/reporting/javascript/app.js --bundle --outfile=mavis/reporting/static/js/app.js --sourcemap --minify --watch"
   },
   "dependencies": {
     "nhsuk-frontend": "^10.0.0"
